hotfix: fix class Card
